event/serializers.py
from rest_framework import serializers
from .models import Event, User


class EventSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Event
        fields = ('title', 'description', 'date', 'place', 'image')



class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ('last_name', 'first_name',)