requirements.txt
asgiref==3.2.6
Django==3.0.4
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.1
Pillow==7.1.1